examples/scaffolds/python/pyproject.toml
[project]
name = "we-plus-matcher-python"
version = "0.1.0"
description = "Python starter for the WE+ x Soulware AI hackathon"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
package = false
